+++ R/setClusters.R
@@ -6,6 +6,8 @@
 #'
 #' @param prefix a prefix to append before the cluster number
 #'
+#' @param ... additional parameters passed to \link{optClust}
+#' 
 #' @return
 #'
 #' Returns a cnr object with cluster membership based on Bray-Curtis
@@ -15,19 +17,29 @@
 #' @importFrom stats cutree
 #' 
 #' @export
-setBrayClusters <- function(cnr, tree.height, prefix = "C") {
+setBrayClusters <- function(cnr, tree.height = NULL, prefix = "C", ...) {
     
     assertthat::assert_that("BrayC" %in% colnames(cnr[["Y"]]) == FALSE,
                             msg = "a cluster column already exists")
+
+    if(is.null(tree.height)) {
+        mocp <- optClust(cnr, ...)
+        tree.height <- min.int <- minimum.intersect(mocp)
+        message("tree.height not set, using minum intersect point of ",
+                tree.height, " as tree.height")
+    } else {
+        tree.height <- tree.height
+    }
     
     BrayC <-  cutree(cnr[["hcdb"]], h = tree.height)
-
+    
     if(!is.null(prefix)) {
         BrayC <- paste0(prefix, BrayC)
     }
     
     cnr[["Y"]]$BrayC <- BrayC
     cnr[["tree.height"]] <- tree.height
-
+    
     return(cnr)
 }
+